CRM_*_Form_*::$_action - Don't redeclare
This property is already declared in the parent class; if upstream makes any
trivial change to it (e.g.  increasing visibility from `protected` to
`public`), then it can break.

For an example script to experiment with PHP property visibility, see:
https://gist.github.com/totten/8a1efedf512add98e3e808796b0eac63

Note: This is a formulaic/untested change since I haven't used the
extension.  Suggested for future-proofing.